Add Table struct to hold Column and Row data together
The row data is now generated on using the Table struct and stored on
it to keep the Column and Row information close. Mulitple rows can be
generated now and translated into a PSQL statement. There are limits to
how large this INSERT statement can be that will have to be dealt with.