Table: Handle any type of PSQL array types and randomly generate JSON
Now that the datatype is used instead of the UDTName, the ARRAY type can
be checked and the fake data generation function can be called
recursively to build an array using the underlying datatype. More work
needs to be done in order to create arrays of different sizes, but at
least the ARRAY type supports any underlying datatype that is already
supported.

Included here is also the change to generate JSON randomly rather than
always using the same string. gofakeit is used to generate the
JSONOptions field so that the JSON itself is randomly generated.