Create Config struct parsed from YAML
Moved most of the options from the command line into a YAML file that
can be read. The path to this YAML file can be passed as an argument.

Eventually, the ability to pass the config will be added back to the
comamand line, but for now, having them parsed from a YAML file makes it
easier to see what the options are during development.

The integration tests have been updated to use a YAML file.