Add `count` argument to `Config.Tables`
`count` has been defaulted to 0 just to make the YAML parsing simple for
now. If the `count` is 0, then the config will be checked to see if it's
been defined there. If `count` is still zero, we'll panic because
there's nothing to do. I'm not entirely sure if I like this behavior.

The nix integration test has also been updated to move the `table` and
`count` arguments to the YAML file.