Generate insert statements for mulitple tables
Removed the `tableName` and `count` arguments to the program. For the
time being, these will be completely defined in the YAML file since the
count is specific to the table.

If `count` is not specified for the table in the YAML file, then we'll
default to creating 10 records. Perhaps the `count` program argument can
be added back in to change the default count.