Expand the `table.Validate` command to check the columns provided
Users will be able to specify the type of data that they want in a
column, but in order to do that, we need to validate that the columns
that the user wants to issue a command against actually exists on the
table. By passing in what the user's commands are, we can check the
table and panic if the column doesn't exist on the table or if the
request isn't supported by gofakeit. Currently in progress to implement
FirstName and LastName.