Allow user to specify different data for 'text' type
Currently supports `FirstName` and `LastName`. When the validation is
performed, a new `CustomData` map is built and stored in the metadata.
This metadata is then passed ot `fakeData` along with the column name so
the map can be checked to see if the user has requested a special data
type.