Table: Make `Validation` function only use one loop
Refactor function to only use one loop instead of two. This isn't a
refactor because we lost some behavior here since the new version
doesn't check if the column exists on the table or not. Since the loop
is based on the column from the table, anything listed in the YAML file
that doesn't exist in the table will simply be ignored.