Refactor: Move handle to DB and FKs to sqldatabase
Refactored the `postgresqldb` package to `sqldatabase` package. The core
struct here keeps a handle to the DB and the FKs fetched from DB
together rather than having the FKs floating out in a map that has to be
passed around.

The `postgresqldb` package may come back as a driver when other
databases are added, but that will host information specific to
connecting and querying a PSQL database.