Refactor: Move `TableColumns` to PostgresqlDriver
The implementation to get the columns were PostgreSQL specific. Similar
to the `foreignkeyrelation` package, `column` and `table` were given
their own package so that the structs can be shared between sqldatabase
and the drivers subpackage.

Additionally, the PostgresqlDriver implementation has been renamed and
moved to a new file in the `drivers` package.

To avoid naming conflicts in `main`, `table` now refers to the package
in `dummy/sqldatabase/table` while `tbl` refers to the local variable in
the main loop.

This is part of a work in progress to move all PostgreSQL related
operations to the PostgresqlDriver.